Revert "Validate visitor el support"
