+++ dev/com.ibm.ws.org.apache.jasper.el.2.2/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2017, 2021 IBM Corporation and others.
+# Copyright (c) 2017 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License v1.0
 # which accompanies this distribution, and is available at
@@ -19,8 +19,6 @@ Export-Package: org.apache.el;thread-context=true, \
    org.apache.el.lang;thread-context=true, \
    org.apache.el.parser, \
    org.apache.el.util
-
-instrument.disabled: true
 
 -buildpath: \
 	com.ibm.ws.org.apache.el:jasper-el;version=2.2,\
