+++ dev/io.openliberty.org.apache.jasper.expressionLanguage.4.0/bnd.bnd
@@ -28,7 +28,7 @@ Import-Package: \
 Require-Capability: osgi.contract;
  filter:="(&(osgi.contract=JavaEL)(version=4.0.0))"
 
-instrument.disabled: true
+instrument.ffdc: false
 
 -buildpath: \
     org.apache.tomcat:tomcat-jasper-el;version=10.0.2,\
